requirements.txt
fastapi
uvicorn
sqlalchemy
qdrant-client
redis
google-generativeai
pypdf
python-multipart
pydantic-settings
python-dotenv
aiofiles
requests
